Project2_streamlit/requirements.txt
pandas
numpy
bs4
streamlit
pyspellchecker
symspellpy
nltk
spacy
gensim
scikit-learn==1.3.2
protobuf==4.23.0
textblob
scipy
wordcloud
pyLDAvis
matplotlib
seaborn
bokeh
tensorflow
transformers